requirements.txt
dearpygui>=1.8.0
python-nmap>=0.7.1
requests>=2.28.0
